EES-1186 Extract InfographicChartForm from ChartConfiguration
